rename: service 파일명 camelCase로 수정 및 경로 수정
