--- src/services/authService.ts
@@ -1,4 +1,0 @@
-// 인증 관련 API 호출
-const authService = () => {};
-
-export default authService;
