+++ src/services/index.ts
@@ -1,2 +1,4 @@
-export { default as authService } from "./authService";
+export * from "./mypageService";
+export * from "./userService";
+export * from "./salaryService";
 export * from "./scheduleService";
