+++ src/services/userService.ts
@@ -13,15 +13,3 @@ export const getUserData = async (uid: string) => {
 		console.log("No such document!");
 	}
 };
-
-export const getSalariesData = async () => {
-	const docRef = doc(db, "user", "user1", "salaries", "8WtOEWQak2wun80XvWOr");
-	const docSnap = await getDoc(docRef);
-	const salariesData = docSnap.data();
-	if (docSnap.exists()) {
-		return salariesData;
-	} else {
-		// doc.data()가 undefined일 경우
-		console.log("No such document!");
-	}
-};
